# backend/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FitMeet",
    "description": "Backend do projeto FitMeet feito como desafio do Bootcamp da Sysmap",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Registrar um novo usuário",
        "tags": ["Autenticação"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "cpf": { "type": "string" },
                  "password": { "type": "string", "format": "password" }
                },
                "required": ["name", "email", "cpf", "password"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cadastrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Usuario criado com sucesso" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro Inesperado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Informe os campos obrigatórios corretamente." }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Email ou Cpf Ja pertence a outro usuario!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "O e-mail ou CPF informado já pertence a outro usuário." }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro Inesperado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Erro Inesperado" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/sign-in": {
      "post": {
        "summary": "Autenticar usuário",
        "tags": ["Autenticação"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "format": "password" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login realizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string" },
                    "id": { "type": "string", "format": "uuid" },
                    "name": { "type": "string" },
                    "email": { "type": "string", "format": "email" },
                    "avatar": { "type": "string" },
                    "xp": { "type": "integer" },
                    "level": { "type": "integer" },
                    "achievements": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": { "type": "string" },
                          "criterion": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Senha incorreta",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Senha incorreta." }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Conta Desativada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Esta conta foi desativada e não pode ser utilizada." }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Usuario não encontrado!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Usuário não encontrado" }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro Inesperado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Erro Inesperado" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}